+++ src/data/reactive/stream-async.js
