+++ src/data/reactive/stream-sync.js
