+++ src/pass/lexer/index.js
@@ -2,8 +2,8 @@
 import type Token from '~/data/pass/lexer'
 import type { StateProcess } from '~/pass/lexer/-state'
 
-import { withIterable as initSyncStream, T as SyncStream } from '~/data/stream-sync'
-import { withGenerator as initAsyncStream, T as AsyncStream } from '~/data/stream-async'
+import { withIterable as initSyncStream, T as SyncStream } from '~/data/reactive/stream-sync'
+import { withGenerator as initAsyncStream, T as AsyncStream } from '~/data/reactive/stream-async'
 import * as tokens from '~/data/pass/lexer'
 import {
   init,
