+++ test/data/reactive/stream-sync.test.js
@@ -1,5 +1,5 @@
 // @flow
-import { withIterable } from '~/data/stream-sync'
+import { withIterable } from '~/data/reactive/stream-sync'
 import { iter as getIter } from '~/util/data'
 
 test('finishes at last character', () => {
